Takeover Ad
This feature was designed as a sales component for the Hubbard
Broadcasting websites.  The file, public js,
is the front-facing javascript file that retrieves the information via
ajax call and renders the takeover Ad. The
takeover will change: The background of the website; the ads on the
left and bottom side ( if it is a full, or
aggressive takeover ); create divs on each side of the site that are
clickable to the clients site.